InstantComm application: Basic functionality
> Login
(JS) AJAX request to PHP file requesting the username from MySQL DBMS

> Register
(JS) After validation, making AJAX request to (PHP) file checking if the username already exists in MySQL DBMS and inserting it if not.

> InstantComm
(JS) AJAX request for new messages in the database; if found, display the ones not yet shown.
Another AJAX request to send the message to the database with the username, content, and date of emission of the message